requirements.txt
asgiref==3.8.1
async-timeout==4.0.3
backports.zoneinfo==0.2.1
certifi==2024.6.2
channels==4.1.0
channels-redis==4.2.0
chardet==3.0.4
Django==4.2.13
django-cors-headers==4.4.0
django-environ==0.11.2
django-modeltranslation==0.19.4
djangorestframework==3.15.2
djangorestframework-simplejwt==5.3.1
googletrans==4.0.0rc1
h11==0.9.0
h2==3.2.0
hpack==3.0.0
hstspreload==2024.6.1
httpcore==0.9.1
httpx==0.13.3
hyperframe==5.2.0
idna==2.10
msgpack==1.0.8
PyJWT==2.8.0
pytz==2024.1
redis==5.0.6
rfc3986==1.5.0
sniffio==1.3.1
sqlparse==0.5.0
typing-extensions==4.12.2
